% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{epd_promoter_epigenetic_features}
\alias{epd_promoter_epigenetic_features}
\title{Epigenetic features at EPD promoters from Roadmap Epigenomics Project}
\format{
A data.frame with 23,208 rows and 878 columns, with rows corresponding to promoters and columns to epigenetic features from the Roadmap Epigenetics Project
}
\source{
Roadmap data was downloaded from \link{https://egg2.wustl.edu/roadmap/data/byFileType/peaks/consolidated/broadPeak/} and promoter values were calculated using the
bigwigaverageoverbed tool from UCSC
}
\usage{
epd_promoter_epigenetic_features
}
\description{
A data.frame with epigenetic features for \code{epd_autosomal_protein_coding_gene_promoters} from the Roadmap Epigenomics project
}
\keyword{datasets}
