% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{epd_promoters}
\alias{epd_promoters}
\title{Promoters of autosomal protein-coding genes}
\format{
A GRanges object with 23,208 ranges with metadata columns giving the name of the promoter and the name of the associated gene
}
\source{
Promoters were downloaded from \link{https://epd.epfl.ch/}
}
\usage{
epd_promoters
}
\description{
A GRanges object containing promoters of autosomal protein-coding genes  from Eukaryotic Promoter Database (EPD).
The 1 KB upstream of the transcription start sites were used to define the promoters.
}
\keyword{datasets}
