% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylDriver.R, R/MethylDriver_package.R
\docType{package}
\name{MethylDriver}
\alias{MethylDriver}
\alias{MethylDriver-package}
\title{Compare methylation change of genomic regions to the change in similar regions}
\usage{
MethylDriver(
  home_region_methyl_change,
  regions_genes,
  closest_regions_list,
  multiple_correction = "fdr",
  q_value_cutoff = 0.05
)
}
\arguments{
\item{home_region_methyl_change}{A named vector of methylation change values for each region, with names indicating the regions that the values refer to}

\item{regions_genes}{= Vector of gene names associated with each region. Can also be a list if there can be more than one gene associated with a region}

\item{closest_regions_list}{A list of the indices if the most similar regions for each region. Should be a list with the indices of the}

\item{multiple_correction}{A multiple testing correction method (default is FDR). Should be one of the choices from p.adjust.methods
Row names of feature table should match names of ranges. If supplied, regions which overlap will be excluded from being considered neighbours.}

\item{q_value_cutoff}{= A significance threshold to use for q-values}
}
\value{
A list with the results of MethylDriver. Contains a vector of significant hypermethylated genes, a vector of significant hypomethylated genes and a data.frame with the name, associated gene, methylation change value, mean of the methylation change values for the most similar regions,
standard deviation of the methylation change values for the most similar regions, z-scores, p-values and q-values for all regions.
}
\description{
#' Uses a table of numerical features for genomic regions of interest to identify the row indices for the n most similar other regions for each region.
Similarity is measured using Euclidean distance calculated using all features. A GRanges object with the genomic coordiantes of the regions can be supplied and if so,
overlapping regions will be excluded from being considered neighbours.

Finds regions where methylation change differs significantly from those of most similar regions, indicating that methylation changes in these regions may be driver events in cancer
}
\author{
Richard Heery
}
