% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{promoter_neighborhoods}
\alias{promoter_neighborhoods}
\title{Most similar 100 promoters for each promoter}
\format{
A named list of length 23,208 where the names of the list elements correspond to the promoter names from \code{epd_autosomal_protein_coding_gene_promoters} and the
values to the indices of the 100 most similar promoters in \code{epd_autosomal_protein_coding_gene_promoters} for that promoter.
}
\usage{
promoter_neighborhoods
}
\description{
A list with the indices of the 100 most similar promoters for each promoter in \code{epd_autosomal_protein_coding_gene_promoters} to be used with \code{MethylDriver}
}
\keyword{datasets}
