% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{tcga_cancer_methylation_change_tables}
\alias{tcga_cancer_methylation_change_tables}
\title{Mean methylation change at EPD promoters in TCGA cancer types}
\format{
A list of length 13 where each element is a table 23,208 rows and a column with the methylation cvhange for each matching normal-tumour pair
}
\usage{
tcga_cancer_methylation_change_tables
}
\description{
A list with the methylation change at each promoter from \code{epd_promoters} in matching normal-tumour samples for 13 TCGA cancer types:
BLCA = Bladder Urothelial Carcinoma, BRCA = Bladder Urothelial Carcinoma, COAD = Colon Adenocarcinoma, ESCA = Esophageal Carcinoma, HNSC = Head and Neck Squamous Cell Carcinoma,
KIRC = Kidney Renal Clear Cell Carcinoma, KIRP = Kidney renal papillary cell carcinoma, LIHC = Hepatocellular Carcinoma, LUAD = Lung Adenocarcinoma,
LUSC = Lung Squamous Cell Carcninoma, PRAD = Prostate Adenocarcinoma, THCA = Thyroid Cancer, UCEC = Uterine Corpus Endometrial Carcinoma
}
\keyword{datasets}
